compare: split redline and help-topic code into included .c files
compare.c had grown to ~1800 lines. Extract two large, self-contained
blocks into separate files that are #included into compare.c, following
the existing compare_*.c pattern (single translation unit, static fns):

  - compare_redline.c: --json-redline output implementation (~460 lines)
  - compare_help.c:    `zsv help compare <topic>` printers (~200 lines)

The #include sites preserve declaration ordering (compare_redline.c is
included after zsv_compare_within_tolerance and the forward declaration
of zsv_compare_collect_row), so this is a pure code move with no
behavioral change. compare.c drops to ~1140 lines.

Verified byte-identical output: all 44 compare/compare-tolerance/
compare-redline/compare-redline-v2 tests pass before and after.
